Fixes issues with pan post channel join.
